WORLD_NOTES_FASTAPI/LandMarks_FinalProject/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
python-multipart
python-dotenv
fastapi-jwt-auth
python-decouple
pydantic[email]
passlib[bcrypt]==1.7.4

